backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.25
pg8000==1.30.5  # Driver puro de Python para PostgreSQL
pydantic==1.10.8
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
email-validator==2.0.0



